+++ Source/Core/DotSpatial.Controls/MapLabelLayer.cs
@@ -170,7 +170,8 @@ namespace DotSpatial.Controls
 
             // Gets the features text and calculate the label size
             string txt = category.CalculateExpression(f.DataRow, selected, f.Fid);
-            if (txt == null) return;
+            if (txt == null)
+                return;
 
             Func<SizeF> labelSize = () => g.MeasureString(txt, CacheList.GetFont(symb));
 
@@ -202,7 +203,8 @@ namespace DotSpatial.Controls
                     {
                         LineString ls = geo.GetGeometryN(n) as LineString;
                         double tempLength = 0;
-                        if (ls != null) tempLength = ls.Length;
+                        if (ls != null)
+                            tempLength = ls.Length;
                         if (longestLine < tempLength)
                         {
                             longestLine = tempLength;
@@ -232,7 +234,8 @@ namespace DotSpatial.Controls
 
             // Gets the features text and calculate the label size
             string txt = category.CalculateExpression(f.DataRow, selected, f.Fid);
-            if (txt == null) return;
+            if (txt == null)
+                return;
             var angle = GetAngleToRotate(symb, f);
             Func<SizeF> labelSize = () => g.MeasureString(txt, CacheList.GetFont(symb));
 
@@ -267,7 +270,8 @@ namespace DotSpatial.Controls
 
             // Gets the features text and calculate the label size
             string txt = category.CalculateExpression(f.DataRow, selected, f.Fid);
-            if (txt == null) return;
+            if (txt == null)
+                return;
             var angle = GetAngleToRotate(symb, f);
             Func<SizeF> labelSize = () => g.MeasureString(txt, CacheList.GetFont(symb));
 
@@ -295,7 +299,8 @@ namespace DotSpatial.Controls
                     for (int n = 0; n < geo.NumGeometries; n++)
                     {
                         Polygon pg = geo.GetGeometryN(n) as Polygon;
-                        if (pg == null) continue;
+                        if (pg == null)
+                            continue;
                         double tempArea = pg.Area;
                         if (largestArea < tempArea)
                         {
@@ -318,7 +323,8 @@ namespace DotSpatial.Controls
         /// will replace content with transparent pixels.</param>
         public void Clear(List<Rectangle> rectangles, Color color)
         {
-            if (BackBuffer == null) return;
+            if (BackBuffer == null)
+                return;
             Graphics g = Graphics.FromImage(BackBuffer);
             foreach (Rectangle r in rectangles)
             {
@@ -353,7 +359,8 @@ namespace DotSpatial.Controls
             for (int chunk = 0; chunk < numChunks; chunk++)
             {
                 int numFeatures = ChunkSize;
-                if (chunk == numChunks - 1) numFeatures = features.Count - (chunk * ChunkSize);
+                if (chunk == numChunks - 1)
+                    numFeatures = features.Count - (chunk * ChunkSize);
                 DrawFeatures(args, features.GetRange(chunk * ChunkSize, numFeatures));
 
                 if (numChunks > 0 && chunk < numChunks - 1)
@@ -385,8 +392,35 @@ namespace DotSpatial.Controls
             for (int chunk = 0; chunk < numChunks; chunk++)
             {
                 int numFeatures = ChunkSize;
-                if (chunk == numChunks - 1) numFeatures = features.Count - (chunk * ChunkSize);
-                DrawFeatures(args, features.GetRange(chunk * ChunkSize, numFeatures));
+                if (chunk == numChunks - 1)
+                    numFeatures = features.Count - (chunk * ChunkSize);
+
+                // without checking vertices for NaN this will casuse some issues when reprojecting to certain map projections like
+                // Mercator and Orthographic. a valid coordinate in one projection could get a NaN in another projection.
+                // DrawFeatures(args, features.GetRange(chunk * ChunkSize, numFeatures));
+
+                // here we will test to see if any of the vertices have a NaN value and if so then we will remove that feature index
+                // from the list. it is essentially a filter so the labeler has a complete set of valid vertices to place a label. we
+                // first cache the indexes with invalid vertices and then after that we will remove them from the master list. I suppose
+                // there might already be a feature property that contains the vertice validity but dont know the codebase well enough to
+                // make that determination. it would save computation if it were available before this routine was reached.
+                var featsValid = features.GetRange(chunk * ChunkSize, numFeatures);
+                var invalidList = new List<int>();
+                foreach (int fid in featsValid)
+                {
+                    var isValidVertices = AreFeatureVerticesValid(FeatureSet.ShapeIndices[fid].Parts);
+                    if (!isValidVertices)
+                    {
+                        invalidList.Add(fid);
+                    }
+                }
+
+                foreach (int invfid in invalidList)
+                {
+                    featsValid.Remove(invfid);
+                }
+
+                DrawFeatures(args, featsValid);
 
                 if (numChunks > 0 && chunk < numChunks - 1)
                 {
@@ -406,7 +440,8 @@ namespace DotSpatial.Controls
         /// <param name="selected">If this is true, nothing is painted, because selected labels get painted together with not selected labels.</param>
         public void DrawRegions(MapArgs args, List<Extent> regions, bool selected)
         {
-            if (FeatureSet == null || selected) return;
+            if (FeatureSet == null || selected)
+                return;
 #if DEBUG
             var sw = new Stopwatch();
             sw.Start();
@@ -466,7 +501,8 @@ namespace DotSpatial.Controls
         public void FinishDrawing()
         {
             OnFinishDrawing();
-            if (Buffer != null && Buffer != BackBuffer) Buffer.Dispose();
+            if (Buffer != null && Buffer != BackBuffer)
+                Buffer.Dispose();
             Buffer = BackBuffer;
             FeatureLayer.Invalidate();
         }
@@ -487,7 +523,8 @@ namespace DotSpatial.Controls
                 g.DrawImageUnscaled(Buffer, 0, 0);
             }
 
-            if (BackBuffer != null && BackBuffer != Buffer) BackBuffer.Dispose();
+            if (BackBuffer != null && BackBuffer != Buffer)
+                BackBuffer.Dispose();
             BackBuffer = backBuffer;
             OnStartDrawing();
         }
@@ -540,7 +577,8 @@ namespace DotSpatial.Controls
         /// <param name="angle">Angle in degree the label gets rotated by.</param>
         private static void CollisionDraw(string txt, Graphics g, ILabelSymbolizer symb, IFeature f, MapArgs e, RectangleF labelBounds, List<RectangleF> existingLabels, float angle)
         {
-            if (labelBounds.IsEmpty || !e.ImageRectangle.IntersectsWith(labelBounds)) return;
+            if (labelBounds.IsEmpty || !e.ImageRectangle.IntersectsWith(labelBounds))
+                return;
             if (symb.PreventCollisions)
             {
                 if (!Collides(labelBounds, existingLabels))
@@ -657,7 +695,8 @@ namespace DotSpatial.Controls
             if (symb.UseLabelAngleField)
             {
                 var angleField = symb.LabelAngleField;
-                if (string.IsNullOrEmpty(angleField)) return 0;
+                if (string.IsNullOrEmpty(angleField))
+                    return 0;
                 return ToSingle(feature.DataRow[angleField]);
             }
 
@@ -666,7 +705,8 @@ namespace DotSpatial.Controls
                 if (lineString is LineString ls1)
                 {
                     var ls = GetSegment(ls1, symb);
-                    if (ls == null) return 0;
+                    if (ls == null)
+                        return 0;
                     if (symb.LineOrientation == LineOrientation.Parallel)
                         return ToSingle(-ls.Angle);
                     return ToSingle(-ls.Angle - 90);
@@ -677,6 +717,32 @@ namespace DotSpatial.Controls
         }
 
         /// <summary>
+        /// Determines if the part range has any invalid vertices for labeling. An invalid vertex is one with either an x or y value of
+        /// Infinity, -Infinity, or NaN.
+        /// </summary>
+        /// <param name="prts">part range</param>
+        /// <returns>True if none of the vertices have an x or value of Infinity, -Infinity, or NaN, False if at least one vertex has an
+        /// x or y value of Infinity, -Infinity, or NaN.</returns>
+        private static bool AreFeatureVerticesValid(List<PartRange> prts)
+        {
+            // cycle through all parts
+            foreach (PartRange prt in prts)
+            {
+                // we only need to find the first vertex that has a non-finite value in x or y. if we find it then the result is that the
+                // part range is not valid. The IsFinite () method returns true if a value is a finite number. Infinite (not finite) values
+                // are Infinity, -Infinity, or NaN
+                var idxInvalid = prt.ToList().FindIndex(v => !double.IsFinite(v.X) || !double.IsFinite(v.Y));
+                if (idxInvalid != -1)
+                {
+                    return false;
+                }
+            }
+
+            // if we get to this line then all vertices are valid so the part range is valid
+            return true;
+        }
+
+        /// <summary>
         /// Gets the segment of the LineString that is used to position and rotate the label.
         /// </summary>
         /// <param name="lineString">LineString to get the segment from.</param>
@@ -684,7 +750,8 @@ namespace DotSpatial.Controls
         /// <returns>Null on unnown LineLabelPlacementMethod else the calculated segment. </returns>
         private static LineSegment GetSegment(LineString lineString, ILabelSymbolizer symb)
         {
-            if (lineString.Coordinates.Length <= 2) return new LineSegment(lineString.StartPoint.Coordinate, lineString.EndPoint.Coordinate);
+            if (lineString.Coordinates.Length <= 2)
+                return new LineSegment(lineString.StartPoint.Coordinate, lineString.EndPoint.Coordinate);
 
             var coords = lineString.Coordinates;
             switch (symb.LineLabelPlacementMethod)
@@ -726,7 +793,8 @@ namespace DotSpatial.Controls
         /// <returns>Empty Rectangle if Coordinate is outside of the drawn extent, otherwise Rectangle needed to draw the label.</returns>
         private static RectangleF PlaceLabel(Coordinate c, MapArgs e, Func<SizeF> labelSize, ILabelSymbolizer symb, double angle)
         {
-            if (!e.GeographicExtents.Intersects(c)) return RectangleF.Empty;
+            if (!e.GeographicExtents.Intersects(c))
+                return RectangleF.Empty;
             var lz = labelSize();
             PointF adjustment = Position(symb, lz);
             RotatePoint(ref adjustment, angle); // rotates the adjustment according to the given angle
@@ -747,10 +815,12 @@ namespace DotSpatial.Controls
         private static RectangleF PlaceLineLabel(Geometry lineString, Func<SizeF> labelSize, MapArgs e, ILabelSymbolizer symb, float angle)
         {
             LineString ls = lineString as LineString;
-            if (ls == null) return Rectangle.Empty;
+            if (ls == null)
+                return Rectangle.Empty;
 
             var ls1 = GetSegment(ls, symb);
-            if (ls1 == null) return Rectangle.Empty;
+            if (ls1 == null)
+                return Rectangle.Empty;
 
             return PlaceLabel(ls1.ToGeometry(Geometry.DefaultFactory).Centroid.Coordinate, e, labelSize, symb, angle);
         }
@@ -773,7 +843,8 @@ namespace DotSpatial.Controls
         private static RectangleF PlacePolygonLabel(Geometry geom, MapArgs e, Func<SizeF> labelSize, ILabelSymbolizer symb, float angle)
         {
             Polygon pg = geom as Polygon;
-            if (pg == null) return RectangleF.Empty;
+            if (pg == null)
+                return RectangleF.Empty;
             Coordinate c = symb.LabelPlacementMethod switch
             {
                 LabelPlacementMethod.Centroid => pg.Centroid.Coordinates[0],
@@ -861,7 +932,8 @@ namespace DotSpatial.Controls
         private void DrawFeatures(MapArgs e, IEnumerable<int> features)
         {
             // Check that exists at least one category with Expression
-            if (Symbology.Categories.All(_ => string.IsNullOrEmpty(_.Expression))) return;
+            if (Symbology.Categories.All(_ => string.IsNullOrEmpty(_.Expression)))
+                return;
 
             Graphics g = e.Device ?? Graphics.FromImage(BackBuffer);
             Matrix origTransform = g.Transform;
@@ -873,7 +945,8 @@ namespace DotSpatial.Controls
             }
 
             FastLabelDrawnState[] drawStates = FastDrawnStates;
-            if (drawStates == null) return;
+            if (drawStates == null)
+                return;
 
             // Sets the graphics objects smoothing modes
             g.TextRenderingHint = TextRenderingHint.AntiAlias;
@@ -902,7 +975,8 @@ namespace DotSpatial.Controls
                 var catFeatures = new List<int>();
                 foreach (int fid in features)
                 {
-                    if (drawStates[fid] == null || drawStates[fid].Category == null) continue;
+                    if (drawStates[fid] == null || drawStates[fid].Category == null)
+                        continue;
                     if (drawStates[fid].Category == category)
                     {
                         catFeatures.Add(fid);
@@ -935,14 +1009,17 @@ namespace DotSpatial.Controls
 
                 foreach (var fid in catFeatures)
                 {
-                    if (!FeatureLayer.DrawnStates[fid].Visible) continue;
+                    if (!FeatureLayer.DrawnStates[fid].Visible)
+                        continue;
                     var feature = FeatureSet.GetFeature(fid);
                     drawFeature(fid, feature);
                 }
             }
 
-            if (e.Device == null) g.Dispose();
-            else g.Transform = origTransform;
+            if (e.Device == null)
+                g.Dispose();
+            else
+                g.Transform = origTransform;
         }
 
         /// <summary>
@@ -953,7 +1030,8 @@ namespace DotSpatial.Controls
         private void DrawFeatures(MapArgs e, IEnumerable<IFeature> features)
         {
             // Check that exists at least one category with Expression
-            if (Symbology.Categories.All(_ => string.IsNullOrEmpty(_.Expression))) return;
+            if (Symbology.Categories.All(_ => string.IsNullOrEmpty(_.Expression)))
+                return;
 
             Graphics g = e.Device ?? Graphics.FromImage(BackBuffer);
             Matrix origTransform = g.Transform;
@@ -966,7 +1044,8 @@ namespace DotSpatial.Controls
             }
 
             Dictionary<IFeature, LabelDrawState> drawStates = DrawnStates;
-            if (drawStates == null) return;
+            if (drawStates == null)
+                return;
 
             // Sets the graphics objects smoothing modes
             g.TextRenderingHint = TextRenderingHint.AntiAlias;
@@ -1029,13 +1108,16 @@ namespace DotSpatial.Controls
 
                 for (int i = 0; i < catFeatures.Count; i++)
                 {
-                    if (!FeatureLayer.DrawnStates[i].Visible) continue;
+                    if (!FeatureLayer.DrawnStates[i].Visible)
+                        continue;
                     drawFeature(catFeatures[i]);
                 }
             }
 
-            if (e.Device == null) g.Dispose();
-            else g.Transform = origTransform;
+            if (e.Device == null)
+                g.Dispose();
+            else
+                g.Transform = origTransform;
         }
 
         #endregion